Rename DB value module
Move the DB value implementation from db.rs to db_value.rs and update internal module paths. Keep the public DB re-export unchanged.
