+++ src/query.rs
@@ -26,7 +26,7 @@ use edn::query::{
 
 use crate::aggregate::{make_accumulator, Accumulator};
 use crate::codec::Decode;
-use crate::db::DB;
+use crate::db_value::DB;
 use crate::expr::{
     expr_variables, BinaryExpr, BinaryOp, Expr, IfExpr, RegexpLikeExpr, UnaryExpr, UnaryOp,
 };
